chore(reduce-imgsize): Reducing go-runner image size
