+++ pkg/log/log.go
@@ -6,8 +6,8 @@ import (
 
 //Fatalf Logs first and then calls `logger.Exit(1)`
 // logging level is set to Panic.
-func Fatalf(msg string, err error) {
-	logrus.WithFields(logrus.Fields{}).Fatalf(msg, err)
+func Fatalf(msg string, err ...interface{}) {
+	logrus.WithFields(logrus.Fields{}).Fatalf(msg, err...)
 }
 
 //Fatal Logs first and then calls `logger.Exit(1)`
